Valider søknad ved innsending
