+++ src/pages/søknad/steg/bo-og-opphold-i-norge/validering.ts
