+++ src/pages/søknad/steg/flyktningstatus-oppholdstillatelse/validering.ts
@@ -1,0 +1,40 @@
+import { SøknadState } from '~src/features/søknad/søknad.slice';
+import { TypeOppholdstillatelse } from '~src/features/søknad/types';
+import { Nullable } from '~src/lib/types';
+import yup from '~src/lib/validering';
+
+export type FormData = SøknadState['flyktningstatus'];
+
+export const schema = yup.object<FormData>({
+    erFlyktning: yup.boolean().nullable().required('Fyll ut om du er registrert flyktning'),
+    erNorskStatsborger: yup.boolean().nullable().required('Fyll ut om du er norsk statsborger'),
+    harOppholdstillatelse: yup
+        .boolean()
+        .nullable(true)
+        .defined()
+        .when('erNorskStatsborger', {
+            is: false,
+            then: yup.boolean().nullable().required('Fyll ut om du har oppholdstillatelse'),
+        }),
+    typeOppholdstillatelse: yup
+        .mixed<Nullable<TypeOppholdstillatelse>>()
+        .nullable(true)
+        .defined()
+        .when('harOppholdstillatelse', {
+            is: true,
+            then: yup
+                .mixed()
+                .nullable()
+                .oneOf(Object.values(TypeOppholdstillatelse), 'Du må velge type oppholdstillatelse')
+                .required(),
+        }),
+    statsborgerskapAndreLand: yup.boolean().nullable().required('Fyll ut om du har statsborgerskap i andre land'),
+    statsborgerskapAndreLandFritekst: yup
+        .string()
+        .nullable(true)
+        .defined()
+        .when('statsborgerskapAndreLand', {
+            is: true,
+            then: yup.string().nullable().min(1).required('Fyll ut land du har statsborgerskap i'),
+        }),
+});
