+++ src/pages/søknad/steg/formue/søkersFormue/DinFormue.tsx
@@ -6,7 +6,7 @@ import { useNavigate } from 'react-router-dom';
 
 import Feiloppsummering from '~src/components/feiloppsummering/Feiloppsummering';
 import { BooleanRadioGroup } from '~src/components/formElements/FormElements';
-import søknadSlice, { Kjøretøy, SøknadState } from '~src/features/søknad/søknad.slice';
+import søknadSlice, { Kjøretøy } from '~src/features/søknad/søknad.slice';
 import SøknadSpørsmålsgruppe from '~src/features/søknad/søknadSpørsmålsgruppe/SøknadSpørsmålsgruppe';
 import { focusAfterTimeout } from '~src/lib/formUtils';
 import { useI18n } from '~src/lib/i18n';
@@ -16,12 +16,10 @@ import { useAppDispatch, useAppSelector } from '~src/redux/Store';
 import Bunnknapper from '../../../bunnknapper/Bunnknapper';
 import * as sharedStyles from '../../../steg-shared.module.less';
 import sharedI18n from '../../steg-shared-i18n';
-import { formueValideringSchema } from '../formueSøknadUtils';
 import KjøretøyInputFelter from '../kjøretøyInputfelter/KjøretøyInputFelter';
+import { FormData, formueValideringSchema } from '../validering';
 
 import messages from './dinformue-nb';
-
-type FormData = SøknadState['formue'];
 
 const DinFormue = (props: { forrigeUrl: string; nesteUrl: string; avbrytUrl: string }) => {
     const formueFraStore = useAppSelector((s) => s.soknad.formue);
