+++ src/pages/søknad/steg/inntekt/søkersInntekt/Inntekt.tsx
@@ -6,7 +6,7 @@ import { useNavigate } from 'react-router-dom';
 
 import Feiloppsummering from '~src/components/feiloppsummering/Feiloppsummering';
 import { BooleanRadioGroup } from '~src/components/formElements/FormElements';
-import søknadSlice, { SøknadState } from '~src/features/søknad/søknad.slice';
+import søknadSlice from '~src/features/søknad/søknad.slice';
 import SøknadSpørsmålsgruppe from '~src/features/søknad/søknadSpørsmålsgruppe/SøknadSpørsmålsgruppe';
 import { focusAfterTimeout } from '~src/lib/formUtils';
 import { MessageFormatter, useI18n } from '~src/lib/i18n';
@@ -17,13 +17,11 @@ import { useAppDispatch, useAppSelector } from '~src/redux/Store';
 import Bunnknapper from '../../../bunnknapper/Bunnknapper';
 import * as sharedStyles from '../../../steg-shared.module.less';
 import sharedI18n from '../../steg-shared-i18n';
-import { inntektsValideringSchema } from '../inntektSøknadUtils';
 import PensjonsInntekter from '../pensonsinntekter/Pensjonsinntekter';
 import TrygdeytelserInputFelter from '../TrygdeytelserInputs/TrygdeytelserInputs';
+import { inntektsValideringSchema, FormData } from '../validering';
 
 import messages from './inntekt-nb';
-
-type FormData = SøknadState['inntekt'];
 
 const DinInntekt = (props: { forrigeUrl: string; nesteUrl: string; avbrytUrl: string }) => {
     const inntektFraStore = useAppSelector((s) => s.soknad.inntekt);
